style : post(게시글) spectacular 문서 분리
