+++ repo/records/posts/urls.py
@@ -4,6 +4,6 @@ from repo.records.posts import views
 
 urlpatterns = [
     path("", views.PostListCreateAPIView.as_view(), name="post-list-create"),
-    path("<int:pk>/", views.PostDetailApiView.as_view(), name="post-detail"),
+    path("<int:pk>/", views.PostDetailAPIView.as_view(), name="post-detail"),
     path("top/", views.TopSubjectPostsAPIView.as_view(), name="post-top"),
 ]
